# server/docs/swagger.yaml
basePath: /v1
definitions:
  domain.Routine:
    properties:
      created_at:
        type: string
      id:
        type: integer
      interval:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  request.CreateRoutineRequest:
    properties:
      interval:
        type: string
      title:
        type: string
    required:
    - interval
    - title
    type: object
  response.APIError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.APIError'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: This is the backend API for the Routine App.
  title: Routine App API
  version: "1.0"
paths:
  /routines:
    get:
      consumes:
      - application/json
      description: Get all routines
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Routine'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Get all routines
      tags:
      - routines
    post:
      consumes:
      - application/json
      description: Create a new routine with the provided title and interval
      parameters:
      - description: Routine to create
        in: body
        name: routine
        required: true
        schema:
          $ref: '#/definitions/request.CreateRoutineRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Routine'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "409":
          description: Conflict
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Create a new routine
      tags:
      - routines
  /routines/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a routine by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Delete a routine
      tags:
      - routines
    get:
      consumes:
      - application/json
      description: Get one routine by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Routine'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Get one routine
      tags:
      - routines
    put:
      consumes:
      - application/json
      description: Update a routine by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Routine'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/response.APIError'
              type: object
      summary: Update a routine
      tags:
      - routines
swagger: "2.0"